refactor: eliminate type duplication between Extension and Web-view
ButtonConfig and RefreshButtonConfig were separately defined in Extension and Web-view
Unified types in src/shared/types.ts for consistency

fix #74
