+++ src/extension/src/types.ts
@@ -1,16 +1,9 @@
-export type ButtonConfig = {
-  color?: string;
-  command?: string;
-  executeAll?: boolean;
-  group?: ButtonConfig[];
-  name: string;
-  shortcut?: string;
-  terminalName?: string;
-  useVsCodeApi?: boolean;
-};
-
-export type RefreshButtonConfig = {
-  color: string;
-  enabled: boolean;
-  icon: string;
-};
+export type {
+  ButtonConfig,
+  RefreshButtonConfig,
+  WebviewMessage,
+  WebviewMessageType,
+  ExtensionMessage,
+  ExtensionMessageType,
+  ConfigurationTarget,
+} from "../../shared/types";
