+++ src/extension/tsconfig.json
@@ -5,7 +5,6 @@
     "outDir": "./out",
     "lib": ["ES2020", "dom"],
     "sourceMap": true,
-    "rootDir": ".",
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
@@ -13,5 +12,6 @@
     "resolveJsonModule": true,
     "types": ["node", "jest", "vscode"]
   },
+  "include": ["src/**/*", "../shared/**/*"],
   "exclude": ["node_modules", ".vscode-test", "coverage"]
 }
